refactor: 기물 별 중복되는 cpature Test를 공통으로 테스트하도록 수정
